requirements.txt
Django ==5.0.4
djangorestframework ==0.1.0
psycopg2-binary